#88 멤버십 설명 삭제 & 멤버십 가입 페이지에 혜택 설명 추가
