+++ src/Components/Header.js
@@ -222,7 +222,7 @@ const Header = () => {
         </Nav>
         {size !== "small" ? (
           <Nav direction='row' className='Menus' gap='large' align='center'>
-            <Link to='/explain' className={isBill && 'displayNone'}> <MemButton>멤버십 가입</MemButton></Link>
+            <Link to='/signIn' className={isBill && 'displayNone'}> <MemButton>멤버십 가입</MemButton></Link>
             <Link to='/brand'>브랜드 소개</Link>
             <span className='DropMenu'>
               인공지능 글쓰기 서비스 <Down size='small' />
@@ -286,7 +286,7 @@ const Header = () => {
           SetShow(false);
         }}>
           <Nav direction='column' className='MobileMenus'>
-            <Link to='/explain' className={isBill && 'displayNone'}>멤버십 가입</Link>
+            <Link to='/signIn' className={isBill && 'displayNone'}>멤버십 가입</Link>
             <Link to='/brand'>브랜드 소개</Link>
             <span className='DropMenu' onClick={HandleMobile}>
               인공지능 글쓰기 서비스 <Down size='small' /></span>
