+++ src/styles/Main.scss
@@ -868,7 +868,25 @@ hr {
 }
 
 .TextCon {
-    padding: 48px 0;
+    padding: 120px 48px;
+    max-width: 1020px;
+
+    > h2 {
+        
+        font-size: 1.2rem;
+        display: flex;
+        flex-direction: column;
+        align-items: center;
+        gap: 15px;
+        font-weight: 600;
+        padding-bottom: 20px;
+        border-bottom : 1px solid $dark-1;
+        margin-bottom : 15px;
+    }
+
+    @media screen and (max-width: 680px) {
+        padding: 100px 20px;
+    }
 }
 
 .ExplainBottom {
@@ -894,6 +912,10 @@ hr {
     > h3 {
         font-size: 1.3rem;
     }
+}
+
+.SignCardGrid {
+    padding: 36px;
 }
 
 .SignInChange {
